backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
distro==1.9.0
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-aiplatform==1.145.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-resource-manager==1.17.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.70.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
requests==2.33.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
